% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abbrev_num}
\alias{abbrev_num}
\title{Returns an abbreviated version of a number (K for thousndas and M for millions)}
\usage{
abbrev_num(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a character
}
\description{
We won't do more than millions of rows for now.
}
\examples{
abbrev_num(1950) # 1.95K
abbrev_num(1950000) # 1.95M

}
