% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{callout_col_labels}
\alias{callout_col_labels}
\title{Callout specification for column labels}
\usage{
callout_col_labels(spec = NULL, color = "black", background = "white")
}
\arguments{
\item{background}{}
}
\value{

}
\description{
Callout specification for column labels
}
