% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{callout_cols}
\alias{callout_cols}
\title{Callout specification for columns}
\usage{
callout_cols(spec = NULL, color = "black", background = "white")
}
\arguments{
\item{background}{}
}
\value{

}
\description{
Callout specification for columns
}
