% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{callout_rows}
\alias{callout_rows}
\title{Callout specification for rows}
\usage{
callout_rows(spec = NULL, color = "black", background = "white")
}
\arguments{
\item{background}{}
}
\value{

}
\description{
Callout specification for rows
}
