% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{callout_text}
\alias{callout_text}
\title{This creates a "callout" object which is essentially a span that has classes, and an id associated
with it, which allows it to be highlighted on a hover.}
\usage{
callout_text(string, which_code_match = 1)
}
\arguments{
\item{string}{the callout string}

\item{which_code_match}{TODO: use this numeric to specify which regex'd string to use if there are multiple
in a given line}
}
\value{

}
\description{
This creates a "callout" object which is essentially a span that has classes, and an id associated
with it, which allows it to be highlighted on a hover.
}
