% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_summary}
\alias{code_summary}
\title{This creates a summary div for the stepper explanation.}
\usage{
code_summary(...)
}
\arguments{
\item{...}{a variable number of `character` or a `callout` object.}
}
\value{
a `[shiny::div(shiny::p(...), class = "summary"), [character]]`
}
\description{
This creates a summary div for the stepper explanation.
}
