% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_text}
\alias{color_text}
\title{Return HTML string(s) containing highlighted text.}
\usage{
color_text(text, type = NULL)
}
\arguments{
\item{text}{a `character` or [`character`] to highlight}

\item{type}{the css class}
}
\value{

}
\description{
Return HTML string(s) containing highlighted text.
}
