% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode_analyzer.R
\name{columns_in_verbs}
\alias{columns_in_verbs}
\title{Given a quoted dplyr code, return a list of <expression, columns used> pairs.}
\usage{
columns_in_verbs(quoted)
}
\arguments{
\item{quoted}{dplyr code}
}
\value{
A list of "expr" / "columns"
  $expr
  [1] "group_by(year, sex)"
  $columns
  text start_col end_col
  1 year        10      13
  2  sex        16      18
}
\description{
Given a quoted dplyr code, return a list of <expression, columns used> pairs.
}
\examples{
require(babynames)
"babynames \%>\%
   group_by(year, sex) \%>\%
   summarise(total = sum(n)) \%>\%
   spread(sex, total) \%>\%
   mutate(percent_male = M / (M + F) * 100, ratio = M / F)" -> pipeline
quoted <- rlang::parse_expr(pipeline)
columns_in_verbs(quoted)
}
