% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{construct_diff_data}
\alias{construct_diff_data}
\title{Takes the diff information from daff and reconstructs its information to consume
later for custom formatting.}
\usage{
construct_diff_data(diff)
}
\arguments{
\item{diff}{}
}
\value{
list with data, and vectors of inserted/modified/deteleted rows/cols
list(
  data = diff_data,
  inserted_rows = inserted_rows,
  inserted_cols = inserted_cols,
  modified_rows = modified_rows,
  deleted_cols = deleted_cols
)
}
\description{
Takes the diff information from daff and reconstructs its information to consume
later for custom formatting.
}
