% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawhats.R
\name{create_group_item_tags}
\alias{create_group_item_tags}
\title{Helper function to create a group item div for SortableJS}
\usage{
create_group_item_tags(lines, ns_id)
}
\arguments{
\item{ns_id}{}
}
\value{

}
\description{
Helper function to create a group item div for SortableJS
}
