% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{daff_html}
\alias{daff_html}
\title{Takes in knitr chunk options and uses the previous dataframe and the
current one to generate a data diff.}
\usage{
daff_html(
  data_ref,
  data,
  include_summary = FALSE,
  include_diff_table = FALSE,
  reactable_output = FALSE
)
}
\arguments{
\item{include_diff_table}{}
}
\value{

}
\description{
Takes in knitr chunk options and uses the previous dataframe and the
current one to generate a data diff.
}
