% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawhats.R
\name{datawatsUI}
\alias{datawatsUI}
\title{Datawats UI}
\usage{
datawatsUI(id, code)
}
\arguments{
\item{id}{A character}

\item{dplyr_code}{A character}
}
\value{



}
\description{
Datawats UI
}
\examples{
Datawats UI

}
