% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{debug_print}
\alias{debug_print}
\title{Debug print if options$debug flag is set to TRUE}
\usage{
debug_print(debug, msg)
}
\arguments{
\item{msg}{}
}
\value{

}
\description{
Debug print if options$debug flag is set to TRUE
}
