% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_indices.R
\name{diet_indices}
\alias{diet_indices}
\title{Computes diet indices: trophic level and omnivory index}
\usage{
diet_indices(Adjacency = NULL, return_TL = TRUE)
}
\arguments{
\item{Adjacency}{The adjacency matrix of the food web. It must be of class matrix.}

\item{return_TL}{whether to return only the omnivory index or both metrics.
Trophic Level is computed in any case, so it is not quicker to set `return_TL` to false.}
}
\value{
A dataframe with two variables, `TL` is the Trophic Level (starting at 0) and `OI` is the omnivory index.
       (if `return_TL` is set to `FALSE` than it's only only column with the Omnivory Index).

 @export
}
\description{
Computes diet indices: trophic level and omnivory index
}
