% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{diff_summary_html}
\alias{diff_summary_html}
\title{Takes a daff summary table to extract pertinent information and returns a
customized table visualizing the changes.}
\usage{
diff_summary_html(s, reactable_output = FALSE)
}
\arguments{
\item{s}{"data_diff_summary"}
}
\value{
knit_print
}
\description{
Takes a daff summary table to extract pertinent information and returns a
customized table visualizing the changes.
}
