% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{explain}
\alias{explain}
\title{Constructor for an explanation of a given line of code}
\usage{
explain(summary = "", ...)
}
\arguments{
\item{summary}{the explanation text}

\item{...}{variable number of `callout`}
}
\value{

}
\description{
Constructor for an explanation of a given line of code
}
