% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{format_columns}
\alias{format_columns}
\title{Takes a dataframe with subset of columns and a daff change type
to format each column appropriate for the given columns.}
\usage{
format_columns(columns, change_type)
}
\arguments{
\item{columns}{data.frame}

\item{change_type}{character}
}
\value{

}
\description{
Takes a dataframe with subset of columns and a daff change type
to format each column appropriate for the given columns.
}
