% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{format_table}
\alias{format_table}
\title{Takes diff information to format and return a reactable table as knit_print.}
\usage{
format_table(diff_data, reactable_output = FALSE)
}
\arguments{
\item{diff}{}
}
\value{
knit_print
}
\description{
Takes diff information to format and return a reactable table as knit_print.
}
