% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_node_.R
\name{fw_node_colour}
\alias{fw_node_colour}
\title{Adds node colour to food web}
\usage{
fw_node_colour(foodweb, colour_fun = tidygraph::centrality_betweenness, ...)

fw_node_colour(foodweb, colour_fun = tidygraph::centrality_betweenness, ...)
}
\arguments{
\item{foodweb}{The food web. It must be in `tbl_graph` format.}

\item{colour_fun}{The **node** function to compute node size (e.g., a centrality metric)}

\item{...}{additional parameters to pass to the function}
}
\value{
a `tbl_graph` with one more variable for the nodes, named `colour`.

a `tbl_graph` with one more variable for the nodes, named `colour_index`.
}
\description{
Adds node colour to food web

Adds node colour to food web
}
\details{
Food web must be in `tbl_graph` format.
         It overrides any other `colour_index` variable if already present for the web.

Food web must be in `tbl_graph` format.
         It overrides any other `colour_index` variable if already present for the web.
}
