% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw_node_.R
\name{fw_node_size}
\alias{fw_node_size}
\title{Adds node size to food web}
\usage{
fw_node_size(foodweb, size_fun = tidygraph::centrality_degree, ...)
}
\arguments{
\item{foodweb}{The food web. It must be in `tbl_graph` format.}

\item{size_fun}{The **node** function to compute node size (e.g., a centrality metric)}

\item{...}{additional parameters to pass to the function}
}
\value{
a `tbl_graph` with one more variable for the nodes, named `size`.
}
\description{
Adds node size to food web
}
\details{
Food web must be in `tbl_graph` format.
         It overrides any other `size` variable if already present for the web.
}
