% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{get_chain_outputs}
\alias{get_chain_outputs}
\title{Given an tidyverse code expression, this will return a list of all of the outputs.
It calls \code{get_output_intermediates} to return a simpler representation of all
of the intermediate outputs. We simply extract the outputs only from the larger
intermediate data structure.}
\usage{
get_chain_outputs(pipeline)
}
\arguments{
\item{pipeline}{tidyverse code expression}
}
\value{
a list of outputs corresponding to each verb in tidyverse code
}
\description{
Given an tidyverse code expression, this will return a list of all of the outputs.
It calls \code{get_output_intermediates} to return a simpler representation of all
of the intermediate outputs. We simply extract the outputs only from the larger
intermediate data structure.
}
\examples{
\dontrun{
require(tidyverse)
quoted <- rlang::parse_expr(
diamonds \%>\%
  select(carat, cut, color, clarity, price) \%>\%
  group_by(color) \%>\%
  summarise(n = n(), price = mean(price)) \%>\%
  arrange(desc(color))
)
outputs <- get_chain_outputs(quoted)
}
}
