% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcode_analyzer.R
\name{get_change_type}
\alias{get_change_type}
\title{Based on the type of tidyr/dplyr function used return whether or not
the type of change was internal (no visible change), visible, or none.}
\usage{
get_change_type(verb_name)
}
\arguments{
\item{verb_name}{}
}
\value{
a character
}
\description{
Based on the type of tidyr/dplyr function used return whether or not
the type of change was internal (no visible change), visible, or none.
}
