% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffing.R
\name{get_daff_output}
\alias{get_daff_output}
\title{Given knitr chunk options for exercise, return the daff output.}
\usage{
get_daff_output(options)
}
\arguments{
\item{options}{}
}
\value{

}
\description{
Given knitr chunk options for exercise, return the daff output.
}
