% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{get_diagnosis}
\alias{get_diagnosis}
\title{Get a diagnosis \code{reactable} given a data.frame/tibble.}
\usage{
get_diagnosis(dat)
}
\arguments{
\item{dat}{\code{data.frame} or \code{tibble}}
}
\value{
\code{reactable}
}
\description{
Get a diagnosis \code{reactable} given a data.frame/tibble.
}
