% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{get_summary}
\alias{get_summary}
\title{Return basic descriptive stats about the columns of a \code{data.frame} or \code{tibble}}
\usage{
get_summary(dat)
}
\arguments{
\item{dat}{\code{data.frame} or \code{tibble}}
}
\value{
\code{tibble}
}
\description{
Currently, we return type of column, # unique elements, # missing values.
}
