% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfoodweb.R
\name{ggfoodweb}
\alias{ggfoodweb}
\title{Plots a foodweb, using and informative layer, node size, and node colour.}
\usage{
ggfoodweb(
  foodweb,
  layer_foodweb,
  dodge = 0,
  max_node_size = 8,
  min_node_size = 4,
  max_edge_width = 1,
  min_edge_width = max_edge_width
)
}
\arguments{
\item{foodweb}{A foodweb in `tbl_graph` or `igraph` format, with *m* species. The node table must have `size` and `colour` variable.}

\item{layer_foodweb}{A *m x 2* matrix. The first column defines the horizontal position and the second column the vertical position of the nodes.}

\item{dodge}{The random uniform noise used to scatter the nodes along the horizontal axis (so to avoid overlap).}

\item{max_node_size}{maximum edge width for plotting}

\item{min_node_size}{minimum edge width for plotting}

\item{max_edge_width}{maximum edge width for plotting}

\item{min_edge_width}{minimum edge width for plotting (default equal to `max_edge_with`, which gives a constant width)}
}
\value{
A `ggraph` plot.
}
\description{
It requires a `foodweb` and a foodweb layer, `layer_foodweb`, providing the vertical and
horizontal positions of the nodes in the plot. A layer like that can be computed using `tl_oi_layer()`.
}
\details{
Moreover, the size and colour of the nodes must be present in the foodweb object.
}
