% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{highlight_code}
\alias{highlight_code}
\title{Color for simple code text highlight.}
\format{
An object of class \code{character} of length 1.
}
\usage{
highlight_code
}
\description{
Color for simple code text highlight.
}
\keyword{datasets}
