% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{knit_print.tutorial_stepper}
\alias{knit_print.tutorial_stepper}
\title{Title}
\usage{
\method{knit_print}{tutorial_stepper}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\value{

}
\description{
Title
}
