% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_content_change}
\alias{log_content_change}
\title{Log a content change from the RStudio editor and Unravel.}
\usage{
log_content_change(content, context = "rstudio")
}
\arguments{
\item{context}{The context}

\item{code}{A character form of the code}
}
\value{
Nothing
}
\description{
Log a content change from the RStudio editor and Unravel.
}
