% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_highlighting.R
\name{prismCodeBlock}
\alias{prismCodeBlock}
\title{Takes in code string and returns a shiny::tagList with a Prism.highlightAll()
script to syntax highlight it.}
\usage{
prismCodeBlock(code_text)
}
\arguments{
\item{code}{}
}
\value{

}
\description{
Takes in code string and returns a shiny::tagList with a Prism.highlightAll()
script to syntax highlight it.
}
