% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{python_df}
\alias{python_df}
\title{Takes in an unconverted pandas.DataFrame via `reticulate::py_eval` and returns
an R data.frame that has the look of a pandas.DataFrame}
\usage{
python_df(pydf)
}
\arguments{
\item{pydf}{}
}
\value{

}
\description{
Takes in an unconverted pandas.DataFrame via `reticulate::py_eval` and returns
an R data.frame that has the look of a pandas.DataFrame
}
