% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reappend}
\alias{reappend}
\title{Takes a number of lists and appends them into a single list.
Instead of doing nested append calls, this will automate that
by doing the series of append calls for you.}
\usage{
reappend(...)
}
\arguments{
\item{...}{a series of lists}
}
\value{
named list
}
\description{
Takes a number of lists and appends them into a single list.
Instead of doing nested append calls, this will automate that
by doing the series of append calls for you.
}
\examples{
reappend(list(a = 1), list(b = 2))

}
