% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{recurse_dplyr}
\alias{recurse_dplyr}
\title{Given a quoted dplyr chain code, return a list of intermediate expressions.
The returned list would include the dataframe name expression.}
\usage{
recurse_dplyr(dplyr_tree, outputs = list())
}
\arguments{
\item{dplyr_tree}{quoted dplyr code}

\item{outputs}{list}
}
\value{
a list
}
\description{
Given a quoted dplyr chain code, return a list of intermediate expressions.
The returned list would include the dataframe name expression.
}
