% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinline_to_html}
\alias{rinline_to_html}
\title{Takes in a markdown string and renders it using `knitr::knit2html`}
\usage{
rinline_to_html(text)
}
\arguments{
\item{text}{}
}
\value{

}
\description{
Takes in a markdown string and renders it using `knitr::knit2html`
}
