% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{stepper_module_ui}
\alias{stepper_module_ui}
\title{Generates the HTML for the stepper UI}
\usage{
stepper_module_ui(id)
}
\value{
list that is a shiny.tag.list containing HTML and other Shiny meta data
}
\description{
Generates the HTML for the stepper UI
}
\examples{
\dontrun{
stepper_module_ui("hello")
}
}
