% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{stepper_prerendered_chunk}
\alias{stepper_prerendered_chunk}
\title{This invokes the stepper Shiny module}
\usage{
stepper_prerendered_chunk(stepper, ...)
}
\arguments{
\item{stepper}{the list structure returned by the `stepper` function}

\item{...}{additional fields}
}
\value{
nothing
}
\description{
This invokes the stepper Shiny module
}
