% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_oi_layer.R
\name{tl_oi_layer}
\alias{tl_oi_layer}
\title{Computes a ggraph layer using the diet indices of the foodweb}
\usage{
tl_oi_layer(foodweb, dodge = 0.08)
}
\arguments{
\item{foodweb}{A foodweb in `tbl_graph` or `igraph` format.}

\item{dodge}{The random uniform noise used to scatter the nodes along the horizontal axis (so to avoid overlap).}
}
\value{
A matrix with two columns (trophic and omnivory) and as many rows as nodes in the foodweb
}
\description{
Computes a ggraph layer using the diet indices of the foodweb
}
\details{
The layer is build by using the Trophic Level as vertical axis and the Omnivory Index as horizontal axis.
}
