% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_simple.R
\name{to_simple}
\alias{to_simple}
\title{A morpher to simplify tidygraphs}
\usage{
to_simple(graph, remove_multiples = TRUE, remove_loops = FALSE)
}
\arguments{
\item{graph}{the graph to be simplifies}

\item{remove_multiples}{whether to remove multiple edges}

\item{remove_loops}{whether to remove loops}
}
\description{
It is a direct rip-off from `tidygraph::to_simple()` and should be used within a `morph` situation.
}
