% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuesdayLakes_1984.R
\docType{data}
\name{TuesdayLakes_1984}
\alias{TuesdayLakes_1984}
\title{The Tuesday Lake food web in 1984}
\description{
The Tuesday Lake food web in 1984
}
\references{
Carpenter SR, Kitchell JF, Hodgson JR, Cochran PA, Elser JJ, Elser MM, Lodge DM, Kretchmer D, He X, von Ende C. Regulation of lake primary productivity by food web structure. Ecology. 1987 Dec;68(6):1863-76.
}
\keyword{data,}
\keyword{food}
\keyword{web}
