% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel}
\alias{unravel}
\title{Unravel dplyr expression for exploration}
\usage{
unravel(code, viewer = T)
}
\arguments{
\item{code}{}
}
\value{

}
\description{
Unravel dplyr expression for exploration
}
\examples{
unravel(
mtcars \%>\%
  head(20) \%>\%
  select(mpg)
)
}
