% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel_addin}
\alias{unravel_addin}
\title{The binding function for unraveling code using the add-in.}
\usage{
unravel_addin()
}
\value{
A shiny app
}
\description{
The binding function for unraveling code using the add-in.
}
