% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel_code}
\alias{unravel_code}
\title{Unravels dplyr code for exploration.}
\usage{
unravel_code(code = "", viewer = T)
}
\arguments{
\item{code}{A character}
}
\value{
A shiny app
}
\description{
Unravels dplyr code for exploration.
}
\examples{
unravel_code("mtcars \%>\% select(cyl, mpg)")
}
