% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel_quoted}
\alias{unravel_quoted}
\title{Unravel dplyr expression for exploration}
\usage{
unravel_quoted(code)
}
\arguments{
\item{code}{}
}
\value{

}
\description{
Unravel dplyr expression for exploration
}
\examples{
unravel_quoted(
mtcars \%>\%
  head(20) \%>\%
  select(mpg)
)
}
