% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel_app.R
\name{unravelServer}
\alias{unravelServer}
\title{Unravel server}
\usage{
unravelServer(id, user_code = NULL)
}
\arguments{
\item{id}{the app \code{character} id}

\item{user_code}{the \code{character} code text}
}
\value{
there are no values returned for this module server function
}
\description{
Unravel server
}
\examples{
\dontrun{
unravelServer("unravel", "mtcars \%>\% select(cyl)")
}
}
