% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel_app.R
\name{unravelUI}
\alias{unravelUI}
\title{Unravel UI}
\usage{
unravelUI(id)
}
\arguments{
\item{id}{A character}

\item{dplyr_code}{A character}
}
\value{



}
\description{
Unravel UI
}
\examples{
Unravel UI

}
