% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_pandas.R
\name{annotate_dataframe}
\alias{annotate_dataframe}
\title{Title}
\usage{
annotate_dataframe(
  df,
  show_rownames,
  rows = list(),
  cols = list(),
  col_labels = list()
)
}
\arguments{
\item{col_labels}{}
}
\value{

}
\description{
Title
}
