% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{color_text}
\alias{color_text}
\title{Highlight text color or its background-color and return an HTML string}
\usage{
color_text(x, type = NULL)
}
\arguments{
\item{background}{}
}
\value{

}
\description{
Highlight text color or its background-color and return an HTML string
}
