% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_pandas.R
\name{kable_pandas}
\alias{kable_pandas}
\title{Takes in an unconverted Python dataframe and outputs a kableExtra table.}
\usage{
kable_pandas(
  df,
  rmd = FALSE,
  show_rownames = FALSE,
  rows = list(),
  cols = list(),
  col_labels = list()
)
}
\arguments{
\item{df}{The unocvertend Python dataframe}

\item{rmd}{This is a flag to render in the RStudio IDE for TRUE, FALSE in a rendered document}

\item{show_rownames}{A flag to indicate whether to show the rownames for MultiIndex dataframes}
}
\value{
A character vector of the table source code (very similar to `kable`)
}
\description{
Takes in an unconverted Python dataframe and outputs a kableExtra table.
}
