% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{knit_print.tutorial_stepper}
\alias{knit_print.tutorial_stepper}
\title{knitr print for stepper}
\usage{
\method{knit_print}{tutorial_stepper}(x, ...)
}
\arguments{
\item{x}{the stepper}

\item{...}{additional fields}
}
\value{

}
\description{
knitr print for stepper
}
