% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_code}
\alias{log_code}
\title{Log a code event}
\usage{
log_code(code, user = "unravel")
}
\arguments{
\item{code}{A code snippet that was executed}

\item{user}{In which context the code was run in ('ide' or 'unravel')}
}
\value{
Nothing
}
\description{
A user either executed some code on the IDE or the Unravel app.
}
