% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_content_change}
\alias{log_content_change}
\title{Log a content change from the RStudio editor and Unravel.}
\usage{
log_content_change(content, path = "", context = "rstudio")
}
\arguments{
\item{content}{The code within editor/Unravel}

\item{path}{The path of the editor}

\item{context}{The context of where the code originates from ("rstudio" or "unravel")}
}
\value{
Nothing
}
\description{
Log a content change from the RStudio editor and Unravel.
}
