% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_foodweb.R
\name{plot_foodweb}
\alias{plot_foodweb}
\title{Draws an opinionated plot of a food web.}
\usage{
plot_foodweb(
  foodweb,
  min_node_size = 4,
  max_node_size = 8,
  max_edge_width = 1,
  min_edge_width = max_edge_width,
  dodge = 0.08
)
}
\arguments{
\item{foodweb}{A foodweb in `tbl_graph` or `igraph` format, with *m* species. The node table must have `size` and `colour` variable.}

\item{min_node_size}{minimum edge width for plotting}

\item{max_node_size}{maximum edge width for plotting}

\item{max_edge_width}{maximum edge width for plotting}

\item{min_edge_width}{minimum edge width for plotting (default equal to `max_edge_with`, which gives a constant width)}

\item{dodge}{The random uniform noise used to scatter the nodes along the horizontal axis (so to avoid overlap).}
}
\value{
A `ggraph` plot.
}
\description{
It requires a `foodweb` in `tbl_graph` format.
If no foodweb layer is defined by the user, `layer_foodweb` is computed using `tl_oi_layer()`.
The vertical position of the species is defined by their trophic level;
the horizontal positions of the species is defined by their omnivory index.
The size and colour of the nodes, if not provided by the user, are respectively
the degree and the betwenness centrality of the nodes.
}
