% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_analysis.R
\name{recurse_dplyr}
\alias{recurse_dplyr}
\title{Given a quoted dplyr chain code, return a list of intermediate expressions, including
the dataframe name expression.}
\usage{
recurse_dplyr(dplyr_tree, outputs = list())
}
\arguments{
\item{dplyr_tree}{quoted dplyr code}

\item{outputs}{list}
}
\value{
[`language`]
}
\description{
Given a quoted dplyr chain code, return a list of intermediate expressions, including
the dataframe name expression.
}
\examples{
recurse_dplyr(rlang::parse_expr("mtcars \%>\% select(cyl)"))
}
