% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinline_to_html}
\alias{rinline_to_html}
\title{Takes in an RMarkdown text as `character`, renders it using `knitr::knit2html`
and finally returns the `shiny::HTML` version.}
\usage{
rinline_to_html(text)
}
\arguments{
\item{text}{}
}
\value{
a `shiny::HTML`
}
\description{
Takes in an RMarkdown text as `character`, renders it using `knitr::knit2html`
and finally returns the `shiny::HTML` version.
}
