% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepper.R
\name{stepper}
\alias{stepper}
\title{Title}
\usage{
stepper(id, explanations = list())
}
\arguments{
\item{explanations}{}
}
\value{

}
\description{
Title
}
\examples{
stepper(
  id = "stepper-ex",
  explanations = list(
    list(line = 1, text = "The first thing we want to do is groupby the CID, FE, and select the FID column."),
    list(line = 2, text = "Then we will count the numbers by the groups.")
    )
 )

}
