% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_button}
\alias{summary_button}
\title{Creates a summary button}
\usage{
summary_button(ns_id, inputId, lineid, change_type, value = 0)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Creates a summary button
}
