% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel}
\alias{unravel}
\title{Unravels dplyr code for exploration.}
\usage{
unravel(code = "")
}
\arguments{
\item{code}{A character}
}
\value{
A shiny app
}
\description{
Unravels dplyr code for exploration.
}
\examples{
unravel("mtcars \%>\% select(cyl, mpg)")
}
