% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unravel.R
\name{unravel}
\alias{unravel}
\title{Unravel dplyr expression for exploration. You can either wrap your dplyr/tidyr code in
a call or pipe the code to unravel.}
\usage{
unravel(code = NULL, viewer = T)
}
\arguments{
\item{code}{a language (dplyr/tidyr code)}

\item{viewer}{a boolean on whether to display unravel in viewer pane
This is `TRUE` by default so it's fluid going between editor and viewer rather than window.}
}
\value{
A shiny app
}
\description{
Unravel dplyr expression for exploration. You can either wrap your dplyr/tidyr code in
a call or pipe the code to unravel.
}
\examples{
unravel(
  mtcars \%>\%
    head(20) \%>\%
    select(mpg)
)
mtcars \%>\%
  head(20) \%>\%
  select(mpg) \%>\%
  unravel()
}
